okay fixed allocation issue and working on check_repeated_meta as it has some issue but the memory is now mostly heap allocated
