+++ simple_helpers.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   simple_helpers.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aalsuwai <aalsuwai@student.42.fr>          +#+  +:+       +#+        */
+/*   By: anasr <anasr@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 17:52:09 by anasr             #+#    #+#             */
-/*   Updated: 2022/03/15 13:39:35 by aalsuwai         ###   ########.fr       */
+/*   Updated: 2022/03/16 15:22:20 by anasr            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,10 +60,11 @@ void	skip_isspaces(int *index, char *input)
 		(*index)++;
 }
 
-void	clear(void)
+int		clear(void)
 {
 	printf("\e[1;1H\e[2J"); //this one doesn't delete history
 	// printf("\033c"); //this one deletes the history of the terminal
+	return (0);
 }
 
 char	*ft_strcpy(char *dst, const char *src)
